// src/app/admission/interfaces/admission.interface.ts
export interface IAdmission {
    readonly cuartel: string;
    readonly sector: string;
    readonly SDP: string;
    readonly abreviaturaCentro: string;
    readonly codigoDelegacionDestino: string;
    readonly nombreDelegacionDestino: string;
    readonly direccionDestino: string;
    readonly codigoEncaminamiento: string;
    readonly grabarEnvio: string;
    readonly numeroEnvio: string;
    readonly comunaDestino: string;
    readonly abreviaturaServicio: string;
    readonly codigoAdmision: string;
}